refactor(cli): split cli.py into a package, one module per help panel
`src/orchestrator/cli.py` was 4,300 lines holding every command surface. It is now
`src/orchestrator/cli/`, cut along the panel boundaries #306 drew:

  _app.py       root Typer app, panel names, command order, global callback
  _common.py    helpers shared by more than one module (_client, _print, _repo_arg, ...)
  start.py      init, doctor, up, models, tui
  understand.py profile, understand, state, audit
  change.py     investigate, design, localize, rca, regression
  build.py      ingest, backlog, openspec
  sdlc.py       the sdlc sub-app
  pkg.py        the pkg sub-app
  media.py      the media sub-app
  registry.py   template, contract, task, mcp, catalog
  __init__.py   imports the modules and mounts the sub-apps on `app`

The move is mechanical: no function body changed, and `--help` for all 67 entries
(root, 26 top-level, 40 subcommands) is byte-identical before and after. The console
entry point `orchestrator.cli:app` is unchanged. Command modules import only from
`_app` and `_common`, never the package root, so there is no import cycle.

Everything that named `cli.py` literally moved with it: the two tests that patch
private helpers now patch the owning module; `scripts/state-numbers.py` and
`scripts/render_architecture_svg.py` count commands across the package (the
architecture SVG and PNG are regenerated because the CLI box now reads `cli/`);
STATE-OF-SPINE's source-module count is 344; CLAUDE.md, ARCHITECTURE.md and SETUP.md
describe the package; and the shadow-guard comment in sdlc/codegen.py no longer
claims a single cli.py exists.
